+++ src/geocpp/subspace.h
@@ -34,7 +34,7 @@ private:
     CMatrix m;
     CVector p;
 public:
-    CSubSpace( unsigned int dim_space =3, unsigned int dim_sub =0 );
+    CSubSpace(size_t dim_space = 3, size_t dim_sub = 0);
     CSubSpace(const CVector &pi, const CMatrix &mi, subspaceflags_t createflags = GEOCPP_FROM_EQS);
     
     /** Copy constructor */
